+++ src/tailchaser/tailer.py
@@ -3,7 +3,7 @@
 .. moduleauthor:: Thanos Vassilakis <thanosv@gmail.com>
 
 """
-import argparse
+
 import binascii
 import glob
 import gzip
@@ -46,7 +46,7 @@ class Tailer(object):
     DONT_BACKFILL = False
     READ_PERIOD = 1
     CLEAR_CHECKPOINT = False
-    READ_PAUSE = 0
+    READ_PAUSE = 1
     TMP_DIR = None
     ONLY_BACKFILL = False
 
@@ -54,8 +54,7 @@ class Tailer(object):
     STOPPED = 'STOPPED'
     RUNNING = 'RUNNING'
     WAITING = 'WAITING'
-    ARGS = ('only_backfill', 'dont_backfill', 'read_period', 'clear_checkpoint',
-            'read_pause', 'temp_dir', 'start_of_record_re')
+    ARGS = ('only_backfill', 'dont_backfill', 'read_period', 'clear_checkpoint', 'read_pause', 'temp_dir')
 
     def __init__(self,
                  only_backfill=ONLY_BACKFILL,
@@ -63,8 +62,7 @@ class Tailer(object):
                  read_period=READ_PERIOD,
                  clear_checkpoint=CLEAR_CHECKPOINT,
                  read_pause=READ_PAUSE,
-                 temp_dir=TMP_DIR,
-                 start_of_record_re=None):
+                 temp_dir=TMP_DIR):
 
         self.config = collections.namedtuple('Args', self.ARGS)
         self.config.dont_backfill = dont_backfill
@@ -73,9 +71,6 @@ class Tailer(object):
         self.config.read_period = read_period
         self.config.read_pause = read_pause
         self.config.temp_dir = temp_dir if temp_dir else tempfile.mkdtemp()
-        if start_of_record_re:
-            self.config.start_of_record_re = re.compile(start_of_record_re)
-            self.read_record = self.read_record_with_regex
         self.state = self.STARTING
         self.stats = collections.Counter()
 
@@ -102,7 +97,6 @@ class Tailer(object):
                     is_backfill_file_info = self.next_to_process(source_pattern, checkpoint)
                     if is_backfill_file_info:
                         is_backfill, file_info = is_backfill_file_info
-                        is_backfill = is_backfill or self.config.only_backfill
                         if file_info:
                             if is_backfill or self.config.only_backfill:
                                 producer = self.backfill(file_info)
@@ -164,8 +158,7 @@ class Tailer(object):
     def backfill(self, file_info):
         file_to_process, (sig, st_mtime, offset) = file_info
         log.debug("backfill: %s %s", file_to_process, offset)
-        copied_file_path = os.path.join(self.config.temp_dir, str(sig))
-        self.copy(file_to_process, copied_file_path)
+        copied_file_path = self.copy(file_to_process, self.config.temp_dir)
         for offset, record in self.process(copied_file_path, sig, st_mtime, offset):
             if not record:
                 break
@@ -192,34 +185,6 @@ class Tailer(object):
             byte_read += len(record)
             yield byte_read, record
 
-    def read_record_with_regex(self, file_to_grep):
-        buff = ''
-        first_record = True
-        e = 0
-        s = 0
-        offset = file_to_grep.tell()
-        while True:
-            data = file_to_grep.read(100000)
-            if not data:
-                break
-            buff += data
-
-            while True:
-                match = self.config.start_of_record_re.search(buff, e - s)
-                if not match:
-                    break
-                s, e = match.span(0)
-                if first_record:
-                    first_record = False
-                    continue
-                offset += len(buff[:s])
-                yield offset, buff[:s]
-                buff = buff[s:]
-                if not buff:
-                    break
-        offset += len(buff)
-        yield offset, buff
-
     @staticmethod
     def make_checkpoint_filename(source_pattern, path=None):
         if not path:
@@ -230,7 +195,7 @@ class Tailer(object):
         return os.path.join(path, os.path.basename(slugify(source_pattern) + '.checkpoint'))
 
     @classmethod
-    def file_opener(cls, file_name, mode='r'):
+    def file_opener(cls, file_name, mode='rb'):
         if file_name.endswith('.gz'):
             log.debug("gzip file: %s", file_name)
             return gzip.open(file_name, mode)
@@ -255,11 +220,15 @@ class Tailer(object):
     #                 return shutil.copyfileobj(src_fh, dst_fh)
     #     return shutil.copy2(src, dst)
 
-    def copy(self, src, dst):
-        log.debug("copying: %s to %s", src, dst)
+    def copy(self, src, temp_dir):
+        log.debug("copying: %s to %s", src, temp_dir)
         with self.file_opener(src, 'rb') as src_fh:
+            sig = str(binascii.crc32(six.b(src_fh.read(SIG_SZ))) & 0xffffffff)
+            src_fh.seek(0, 0)
+            dst = os.path.join(temp_dir, sig)
             with open(dst, 'wb') as dst_fh:
-                return shutil.copyfileobj(src_fh, dst_fh)
+                shutil.copyfileobj(src_fh, dst_fh)
+                return dst
 
     @classmethod
     def make_sig(cls, file_to_check, stat=None):
@@ -279,47 +248,3 @@ class Tailer(object):
     def save_checkpoint(self, checkpoint):
         log.debug('dumping %s %s', self.config.checkpoint_filename, checkpoint)
         return pickle.dump(checkpoint, open(self.config.checkpoint_filename, 'wb'))
-
-    @classmethod
-    def build_arg_parser(cls, parser=None):
-        if not parser:
-            parser = argparse.ArgumentParser(description='Process some integers.')
-        parser.add_argument('file-pattern',
-                            help='file pattern to tail, such as /var/log/access.*')
-        parser.add_argument('--only-backfill', action='store_true', default=Tailer.ONLY_BACKFILL,
-                            help='dont\'t tail, default: %s' % Tailer.ONLY_BACKFILL)
-        parser.add_argument('--dont-backfill', action='store_true', default=Tailer.DONT_BACKFILL,
-                            help='basically only tail, default: %s' % Tailer.DONT_BACKFILL)
-        parser.add_argument('--clear-checkpoint', action='store_true', default=Tailer.CLEAR_CHECKPOINT,
-                            help='start form the beginning, default: %s' % Tailer.CLEAR_CHECKPOINT)
-        parser.add_argument('--read-period', type=int, default=Tailer.READ_PERIOD,
-                            help='how long you read before you pause.' +
-                                 'If zero you don\'t pause, default: %s' % Tailer.READ_PERIOD)
-        parser.add_argument('--read-pause', type=int, default=Tailer.READ_PAUSE,
-                            help='how long you pause between reads, default: %s' % Tailer.READ_PAUSE)
-        parser.add_argument('--reading-from', choices=['unix', 'win'], default='win',
-                            help='sets how long you rad and then pause, default: win')
-        parser.add_argument('--temp-dir', default=Tailer.TMP_DIR,
-                            help='on backfil files are copied to a temp directory.' +
-                                 'Use this to set this directory, default: %s' % Tailer.TMP_DIR)
-        parser.add_argument('--logging', choices=['DEBUG', 'INFO', 'WARN', 'ERROR', 'CRITICAL'], default='ERROR',
-                            help='logging level, default: ERROR')
-        parser.add_argument('--start-of-record-re', default=None,
-                            help='use this regex expresion to define the start of a record, default: None')
-        return parser
-
-    @classmethod
-    def cli(cls, argv=sys.argv, parser=None, consumer=None):
-        arg_parser = cls.build_arg_parser(parser)
-        args = vars(arg_parser.parse_args(argv[1:]))
-        file_pattern = args.pop('file-pattern')
-        reading_from = args.pop('reading_from')
-        logging_level = getattr(logging, args.pop('logging'))
-        if reading_from == 'win':
-            args['read_pause'] = 1
-            args['read_period'] = 1
-        else:
-            args['read_pause'] = 0
-            args['read_period'] = 0
-        logging.basicConfig(format='%(asctime)s %(levelname)s:%(message)s', level=logging_level)
-        return Tailer(**args).run(file_pattern, consumer)