+++ CHANGELOG.rst
@@ -3,6 +3,14 @@ Changelog
 =========
 
 0.1.0 (2016-02-21)
------------------------------------------
+------------------
 
 * First release on PyPI.
+
+0.1.2 (2016-04-05)
+--------------------
+
+* Add multi-line record support -  #3
+* add --start-of-record-re option for multi-line records. - #4
+* move argparser definition  from cli to Tailer class. - #5
+